# swagger.yaml
openapi: 3.0.0
info:
  title: Bank of Checkmarx API
  description: API for the Bank of Checkmarx demo application
  version: 1.0.0
  contact:
    name: Bank of Checkmarx
    email: support@bankofcheckmarx.com

servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /api/v1/health:
    get:
      summary: Health check
      description: Check if the API is running
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"

  /api/v1/accounts:
    get:
      summary: Get all accounts
      description: Retrieve all bank accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized

  /api/v1/accounts/{account_id}:
    get:
      summary: Get account by ID
      description: Retrieve a specific account by ID
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found

  /api/v1/transactions:
    get:
      summary: Get all transactions
      description: Retrieve all transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'

  /api/v1/transactions:
    post:
      summary: Create transaction
      description: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Account:
      type: object
      properties:
        id:
          type: string
        account_number:
          type: string
        balance:
          type: number
        currency:
          type: string
        account_type:
          type: string
        created_at:
          type: string
          format: date-time

    Transaction:
      type: object
      properties:
        id:
          type: string
        from_account:
          type: string
        to_account:
          type: string
        amount:
          type: number
        currency:
          type: string
        transaction_type:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time

    TransactionRequest:
      type: object
      required:
        - from_account
        - to_account
        - amount
        - currency
      properties:
        from_account:
          type: string
        to_account:
          type: string
        amount:
          type: number
        currency:
          type: string
        transaction_type:
          type: string
          default: "transfer" 